+++ lib/pages/home_page.dart
@@ -1,3 +1,4 @@
+import 'package:festival_flutterturkiye_org/event_flow/event_flow_section.dart';
 import 'package:flutter/material.dart';
 
 import 'package:festival_flutterturkiye_org/core/model/navigation_action.dart';
@@ -72,6 +73,7 @@ class _HomePageState extends State<HomePage> {
               controller: _scrollController,
               children: const <Widget>[
                 CountdownSection(),
+                EventFlowSection(),
                 FooterSection(),
               ],
             ),
